backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
mcp>=1.0.0
python-dotenv>=1.0.0
setuptools==69.5.1
langtrace-python-sdk>=3.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
